+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 }
 
 group = "com.github.shichuanyes"
-version = "0.9.2"
+version = "0.9.3"
 
 repositories {
     mavenLocal()
