+++ src/main/kotlin/PluginData.kt
@@ -8,9 +8,6 @@ object PluginData : AutoSavePluginData("data") {
     val groupSubscribers: MutableSet<Long> by value()
 
     var lastDynamic: Long by value(0L)
-    // TODO: fix dynamic conflict
-
     val liveStatus: MutableMap<Long, Int> by value()
-
     val watchlist: MutableSet<Long> by value()
 }
