+++ src/main/kotlin/PluginMain.kt
@@ -14,7 +14,7 @@ object PluginMain : KotlinPlugin(
     JvmPluginDescription(
         id = "com.github.shichuanyes.asoul-bot",
         name = "asoul-bot",
-        version = "0.9.2"
+        version = "0.9.3"
     ) {
         author("shichuanyes")
     }
@@ -32,8 +32,10 @@ object PluginMain : KotlinPlugin(
         AbstractPermitteeId.AnyContact.permit(Asoul.permission)
 
         // TODO: use separate periods and add support for changing period
-        Timer("UpdateDynamic", false).schedule(UpdateDynamic, PluginConfig.period / 2, PluginConfig.period)
-        Timer("UpdateLiveStatus", false).schedule(UpdateLiveStatus, PluginConfig.period, PluginConfig.period)
+        Timer("UpdateDynamic", false)
+            .scheduleAtFixedRate(UpdateDynamic, PluginConfig.period, PluginConfig.period)
+        Timer("UpdateLiveStatus", false)
+            .scheduleAtFixedRate(UpdateLiveStatus, PluginConfig.period, PluginConfig.period)
 
         logger.info { "Plugin asoul-bot loaded" }
     }
