+++ src/main/kotlin/Utils.kt
@@ -1,5 +1,6 @@
 package com.github.shichuanyes.plugin.asoul
 
+import kotlinx.coroutines.launch
 import net.mamoe.mirai.Bot
 import net.mamoe.mirai.contact.Contact
 import net.mamoe.mirai.contact.Contact.Companion.uploadImage
@@ -9,9 +10,13 @@ import net.mamoe.mirai.message.data.isContentEmpty
 import java.io.File
 
 object Utils {
-    suspend fun broadcastTextWithImages(bot: Bot, text: PlainText, images: MutableList<File>?) {
-        for (subscriber in PluginData.userSubscribers) sendTextWithImages(bot.getFriend(subscriber), text, images)
-        for (subscriber in PluginData.groupSubscribers) sendTextWithImages(bot.getGroup(subscriber), text, images)
+    suspend fun broadcastTextWithImages(text: PlainText, images: MutableList<File>?) {
+        for (bot in Bot.instances) {
+            PluginMain.launch {
+                for (subscriber in PluginData.userSubscribers) sendTextWithImages(bot.getFriend(subscriber), text, images)
+                for (subscriber in PluginData.groupSubscribers) sendTextWithImages(bot.getGroup(subscriber), text, images)
+            }
+        }
     }
 
     private suspend fun sendTextWithImages(target: Contact?, text: PlainText, images: MutableList<File>?) {
